Rudimentary camera scrolling

+++ core/src/com/afyber/game/api/overworld/CameraLimit.java
@@ -15,7 +15,7 @@ public class CameraLimit extends WorldObject {
     }
 
     @Override
-    public void render(SpriteBatch batch, int cameraX, int cameraY) {
+    public void render(SpriteBatch batch, float cameraX, float cameraY) {
         // none, should be invisible
     }
 }
